refactor: rename common config to client
