+++ src/common/index.ts
@@ -1,1 +1,1 @@
-export * from './config';
+export * from './client';
